chore: move new tests in a new test file
